Fix configure.ac build environment variable `ARCH_CFLAGS`
This var was introduced to allow for architecture specific switches
required for some arm boards.

However there was no AC_SUBST() macro exporting the var and recent tests
where I used the var to pass args for cross-building, showed that it was
not being substituted in Makefile.inc.in, so having no effect.

AC_SUBST(ARCH_CFLAGS) inserted and fixes.

Signed-off-by: Mick <arceye@mgware.co.uk>