hal_temp_bbb.py:  CRAMPS pull-up resistor configurable
In `adc2r_cramps()`, add an `R_PU` arg so the function is reusable for
circuits that use the same voltage divider configuration as the CRAMPS
but with different component values.  Plumb this to a new
`hal_temp_bbb.py --r_pu VAL` command-line argument.

Also adapt @cdsteinkuehler's excellent description of the circuit [1]
for the comments.

[1]: https://groups.google.com/d/msg/machinekit/wZ8KAKqV7yo/blG1yZ0rBAAJ