jplan.c: introduce homing of planner
this commit introduces 2 new pins
home-pos
home-set

when home-set is high, the curr-pos value
will be set to the value of home-pos.

This will enable aligning the planner so
that when the planer is enabled, a move
is planned taking this new curr-pos into
account.

Please note that changing this home-pos
(when home-set is made high) will also
directly change the curr-pos output pin.