limit3 comp:  fix 1st order limit violations and drift
This is a rework done for LCNC a year or so ago.  It keeps track of
the input signal and the min/max limits separately so that when the
input signal exceeds a limit, the output doesn't crash into the limit
trying to follow it.  It also tries to be a little smarter by looking
at the input signal velocity to predict when it will come back in
limit, and starts accelerating early to meet it.

It also fixes the problem where output can drift away from input at
zero velocity, fixed separately by @machinekoder back in July.

See related discussion at these issues & PRs:

https://github.com/LinuxCNC/linuxcnc/issues/240
https://github.com/LinuxCNC/linuxcnc/pull/351
https://github.com/LinuxCNC/linuxcnc/issues/378
https://github.com/LinuxCNC/linuxcnc/pull/396
https://github.com/machinekit/machinekit/pull/1384