Add specific -lzmq to CFLAGS to prevent DSO errors in later distros.
These errors occur with later compilers where despite linkage to a
lib, it is not sequenced to the liking of the compiler and throws
unknown symbol errors.

Has been fixed for comp too in upcoming commit which is part of the
single-module-dir work, but this will get the packages to build
on later compilers.

Signed-off-by: Mick <arceye@mgware.co.uk>