rtapi/flavor:  Fix Xenomai load on non-Xenomai hosts
In previous tests, I somehow missed that when the Xenomai libs are
linked to an executable (or a `dlopen()`ed plugin), if they fail to
detect Xenomai on the running system, they will exit the application
with this message:

    Xenomai: /dev/rtheap is missing
    (chardev, major=10 minor=254)

This seems unavoidable, because this happens in `ld-linux.so` before a
single line of application code is executed, and the [Xenomai code][1]
appears not to provide a way to turn the check off.

This patch works around the problem by splitting `xenomai.c` into two.

`xenomai.c`:
- Still linked to `rtapi.so`
- Retains the flavor descriptor struct
- Retains the flavor checks and module init and exit hooks
- Module init now:
  - Calls `dlopen("xenomai_lodaer.so")`,
  - Extracts the `xenomai_descriptor_updater()` function, and
  - Calls that, passing the flavor descriptor to be updated

`xenomai_loader.c`:
- Linked into separate module with Xenomai libraries
- Contains all functions that depend on Xenomai libs
- Contains a `xenomai_descriptor_updater()` function that plugs those
  functions into the flavor descriptor

In this way, the Xenomai libraries are only loaded when the Xenomai
flavor is picked, and don't exit `rtapi_app` on non-Xenomai systems.

[1]:  https://gitlab.denx.de/Xenomai/xenomai/blob/eol/v2.6.x/src/skins/common/bind.c#L67-72