Makefile:  Revert "** configure.ac has changed **" behavior
This was added to fix a dependency problem in the `Makefile`
introduced in the Universal Build.  That could have been caused by the
recursive make that built the multiple module flavors, or it could
have been caused somewhere else.  Either way, hopefully the problem is
gone now with the single module build changes, and if not, the problem
should be fixed properly.