Build `libhalcmd.so` library
The former TCL `hal.so` module is a binding for `halcmd` commands.
There are major differences in the MK implementation added since the
LCNC fork, even though the API hasn't changed significantly.

The TCL `hal.so` module now needs to be built on the LCNC side, since
TCL has been removed from MK.  In order to make the `halcmd`
implementation available to LCNC, the sources shared by `halcmd` and
`hal.so` are now built into the `libhalcmd.so` shard library.