FEATURE: Pass '-Werror' as DPKG_CFLAGS variable in case GCC is version 7 or lower
Implements functionality which was originally passed externally to Docker image in 'scripts/build_docker' script directly to the image. The DPKG_CFLAGS variable is exported during ENTRYPOINT run in case that the GCC is the right version.

Checking uses the $(CC) variable which will be used for the actual building of packages.