MAINTENANCE: Add compatibility level to rtapi_string used in EMCApplication
EMCApplicaqtion - given how the split between HAL and Featured Applications is now done - needs symbols on the Machinekit-HAL side. There are several ways how to satisfy this need. Either add these patched to Machinekit-HAL (which is what this commit does) or implement a compatibility layer in EMCApplication source tree.

From long term perspective, it is probably better to add the compatibility layer to EMCApplication and do not clutter Machinekit-HAL code with specific changes applicable to only one Featured Application. However, this change is small and so far there is no standard how to go around adding the compatibility layers to EMCApplication, so hopefully it will not cause a big problem. Moreover, this change (commit) can be dropped in the future.