GH workflows:  Make name of runs readable in webpage
The names of the runs are so long that on the GH actions "runs" page,
list items on the left are truncated, with the interesting ones
reading, "Build Machinekit HAL packages ..." and "Test Machinekit HAL
on D...".

The interesting information is the type of run ("Build", "Test"), the
OS vendor ("Debian", "Ubuntu"), the code name ("Stretch", "Focal"),
and the architecture ("amd64", "armhf").

This patch shortens the run descriptions to remove unneeded
information so that the important information is visible.