FEATURE: Add querying Docker image manifest and implement logic for forcing rebuild
This should solve issue reported in machinekit/machinekit-hal#293

Original Github Actions CI workflow was designed for the single branch development style. The caching of Docker image set of builders did not take into account that different development branches. This caused a problem in case that two branches were so fundamentally different that the same builder (primarily because of difference in dependencies) cannot be used, but latest push did not touch the watched files (and by that cause a rebuild), then the build could fail even through it should not.

This commit implements the following logic:

Stop all workflow running at the specific branch
Download Docker image manifests from Docker registry
Query all image's manifests digest SHA and commit SHA label
Test if all images are present (if not, force rebuild) and if all are of the same git commit SHA (if not, force rebuild)
Test if between the git commit SHA from labels and HEAD was any change on watched files (if so, force rebuild)
Store the information in output JSON (so following jobs can query it)
In each parallel (matrix) runner, after docker pull check the digest SHA (if these are not the same, force a failure of the job)

This logic workflow should be immune to multi-branches environment and to force-pushes.