cython:  Fix `del signals['foo']` construction
Deleting from the dict-like `hal_objectdict` doesn't actually do
anything, and the signal still shows up in the `signals` list.

Add a `delete()` method to the `HALObject` base class that's always
called during a `del HALObjectDict()['foo']`; in the
`Signal` subclass, the `delete()` method now deletes the underlying
HAL object.  This replaces old, commented-out code that did the same
thing for `Signal`, but in a hackish way in the base class.

The old `Group.delete()` method had to be moved out of the way to
`member_delete()` for consistency, and for more consistency,
`Group.add()` is renamed to `member_add()`.