Disable syslog_async throttling
`syslog_async` throttles bursts of log output.  When running
`DEBUG=5`, it will slow to the point of printing one line per second,
and many logs are dropped, including important error messages.

This patch simply disabled the throttling; messages may still be
dropped, but far fewer than WITH the throttling.

This is a known problem for a long time, most recently mentioned in an
unrelated issue where it caused problems debugging:

https://github.com/machinekit/machinekit-hal/issues/173#issuecomment-685255302