cython:  Move test sighandlers to pytest fixture
When realtime stops, it may try to kill HAL python processes, which
kills the unit tests.  The old code pasted a signal handler at the end
of each file.

Move the handler to a fixture with comments so that people don't have
to remember to add the sighandler, so they don't have to see the
eyesore, and so they know what it is when they see it in the fixtures.