MAINTENANCE: Make sure the flavor loading function pointers in 'rtapi_app' are coming from the 'rtapi' HAL module
Previously the flavor registering function pointers were not necessarily 'dlsym'ed from the 'RTAPIMOD' loadable module as the Module object on which the '.sym()' method was called was empty. Thus, the symbol resolution was probably happening from global space (modules are loaded 'RTLD_GLOBAL | RTLD_NOW') and not specifically from the wanted shared library.