FEATURE: Implement basic testing for CMake build-system
CMake build-system is able to run the 'runtests' from the BUILD tree. Only the acceptance type of tests are working. All which require compilation and thus probably should not be part of 'runtests', but some other test suite, are currently not working.

Not working are also the Python 'halmodule' test and the 'hostmot2' test.

At the moment, 8 tests fail.