FEATURE: Implement changes required by the 'pre-commit/pre-commit-hooks'
Commits the automatic and manual changes required by the 'pre-commit' configuration for the 'github.com/pre-commit/pre-commit-hooks' repository community hooks as specified in the '.pre-commit-config.yaml' configuration file in the root of Machinekit-HAL code source-tree.

Mostly removal of white space, adding execution bits to files which have a shebang and also removal or shebang from files which should not be executable.