FEATURE: Implement CIRCLE CI testing for 'arm64' architecture
This implements the Circle CI workflow for testing of Machinekit-HAL codebase. The logic is based on workflow ('debian-builder-workflow.yaml') from current GitHub Actions implementation and fully mirrors. (Building packages for all allowed combinations of Debian-like distributions, versions and target architectures, testing these packages where applicable and running the testsuites [Runtests and PyTests] from the CMake build tree.)

Because the CIRCLE CI offers "only" 400000 credits per month to Open-Source projects, the jobs run only on the 'arm64' executors (the 'arm.medium' resource class to be specific), but there is provision to run the tests on the 'amd64' executors too, should that need ever arise in the future. (Thus it will then run the same tests as the current GitHub Actions workflow, in addition to the 'arm64' ones.)