MAINTENANCE: Add CIRCLE CI workflow status button to the README
Present in the README page (main visible page on the GitHub repository landing page) the status of Circle CI workflow (specifically for the https://github.com/machinekit/machinekit-hal repository and the master branch there). And thus advertise to the world that the project is build using the Circle CI service.